Directive string support: Handle update case
The tippy directive properly supports string values on first render.
However, when the component is updated, there was no string
handling and the value is assumed to be an object.

Now we apply the same handling for the update case, too.

It would otherwise fail with something like this:
Error: [tippy]: `0` is not a valid option
    validateOptions
    validateOptions
    set
    componentUpdated
    ...